fixed scene loading glitch
unity decided to set the timescale to 0 some reason, so i just said dont and now it works? there is a visual glitch though still, but that is fixed by just loading the scene twice. then it works?? not a critical issue, but sure is really annoying and needs to be fixed later